+++ app/models/topic.rb
@@ -6,41 +6,75 @@ class Topic < ActiveRecord::Base
 
   # Concerns
   include Deletable
+  include PgSearch
+  multisearchable against: [:title],
+                  unless: :deleted?
 
   # Callbacks
-  after_commit :create_first_comment, on: :create
+  after_commit :create_first_reply, on: :create
 
   # Named Scopes
   scope :not_banned, -> { joins(:user).merge(User.where(banned: false)) }
-  def self.search(arg)
-    # TODO: Implement full text search through PgSearch multisearchable
-    where('1 = 1')
-  end
+  scope :reply_count, -> { select('topics.*, COUNT(replies.id) reply_count').joins(:replies).group('topics.id') }
 
   # Model Validation
   validates :title, :slug, :user_id, presence: true
   validates :title, length: { maximum: 40 }
   validates :description, presence: true, if: :new_record?
-  validates :slug, uniqueness: { scope: :deleted }
+  validates :slug, uniqueness: { scope: :deleted } # TODO: Uniqueness can't be constrained to deleted topics
   validates :slug, format: { with: /\A(?!\Anew\Z)[a-z][a-z0-9\-]*\Z/ }
 
   # Model Relationships
   belongs_to :user
-  has_many :comments, -> { order :id }
+  has_many :topic_users
+  has_many :replies, -> { order :id }
+  has_many :reply_users
   has_many :subscriptions
   has_many :subscribers, -> { current.where(emails_enabled: true).where(subscriptions: { subscribed: true }) }, through: :subscriptions, source: :user
   has_many :topic_tags
-  has_many :tags, -> { where(deleted: false).order(:name) }, through: :topic_tags
+  has_many :tags, -> { current.order(:name) }, through: :topic_tags
+
+  # Model Methods
+  def destroy
+    super
+    update_pg_search_document
+    replies.each(&:update_pg_search_document)
+  end
 
   def to_param
     slug_was.to_s
+  end
+
+  def started_reading?(current_user)
+    topic_user = topic_users.find_by user: current_user
+    topic_user ? true : false
+  end
+
+  def unread_replies(current_user)
+    topic_user = topic_users.find_by user: current_user
+    if topic_user
+      root_replies.current.where('id > ?', topic_user.current_reply_read_id).count
+    else
+      0
+    end
+  end
+
+  def next_unread_reply(current_user)
+    topic_user = topic_users.find_by user: current_user
+    root_replies.current.find_by('id > ?', topic_user.current_reply_read_id) if topic_user
+  end
+
+  def root_replies
+    replies.where(reply_id: nil)
   end
 
   def editable_by?(current_user)
     !locked? && !user.banned? && (user == current_user || current_user.system_admin?)
   end
 
-  # Placeholder
+  def last_page
+    ((replies.where(reply_id: nil).count - 1) / Reply::REPLIES_PER_PAGE) + 1
+  end
 
   def get_or_create_subscription(current_user)
     current_user.subscriptions.where(topic_id: id).first_or_create
@@ -56,8 +90,8 @@ class Topic < ActiveRecord::Base
 
   private
 
-  def create_first_comment
-    comments.create description: description, user_id: user_id if description.present?
+  def create_first_reply
+    replies.create description: description, user_id: user_id if description.present?
     get_or_create_subscription(user)
   end
 end